+++ src/Renderer/VertexBuffer.h
@@ -14,7 +14,7 @@ public:
   VertexBuffer();
 
   void fillBuffer(std::vector<float>& data, GLenum usage = GL_STATIC_DRAW);
-  void fillBuffer(long size, GLenum usage = GL_STATIC_DRAW);
+  void fillBuffer(signed long long int size, GLenum usage = GL_STATIC_DRAW);
   void fillBufferSubData(std::vector<float>& data, int offset = 0);
   void bind();
   void unBind();
